ts debug log
